requirements.txt
flask
pymongo
Pillow